# virsh-sandbox/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing virtual machine sandboxes using libvirt",
        "title": "virsh-sandbox API",
        "contact": {},
        "version": "0.0.1-beta"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/sandbox/create": {
            "post": {
                "description": "Creates a new virtual machine sandbox by cloning from an existing VM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Create a new sandbox",
                "parameters": [
                    {
                        "description": "Sandbox creation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.createSandboxRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.createSandboxResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}": {
            "delete": {
                "description": "Destroys the sandbox and cleans up resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Destroy sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/diff": {
            "post": {
                "description": "Computes differences between two snapshots",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Diff snapshots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Diff parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.diffRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.diffResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/generate/{tool}": {
            "post": {
                "description": "Generates Ansible or Puppet configuration from sandbox changes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Generate configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tool type (ansible or puppet)",
                        "name": "tool",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.generateResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/publish": {
            "post": {
                "description": "Publishes sandbox changes to GitOps repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Publish changes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Publish parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.publishRequest"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.publishResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/run": {
            "post": {
                "description": "Executes a command inside the sandbox via SSH",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Run command in sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Command execution parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.runCommandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.runCommandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/snapshot": {
            "post": {
                "description": "Creates a snapshot of the sandbox",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Create snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Snapshot parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.snapshotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.snapshotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/sshkey": {
            "post": {
                "description": "Injects a public SSH key for a user in the sandbox",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Inject SSH key into sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SSH key injection parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.injectSSHKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sandbox/{id}/start": {
            "post": {
                "description": "Starts the virtual machine sandbox",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandbox"
                ],
                "summary": "Start sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Start parameters",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.startSandboxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.startSandboxResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/vms": {
            "get": {
                "description": "Returns a list of all virtual machines from the libvirt instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMs"
                ],
                "summary": "List all VMs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.listVMsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_rest.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "virsh-sandbox_internal_rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "virsh-sandbox_internal_rest.createSandboxRequest": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "description": "required",
                    "type": "string"
                },
                "cpu": {
                    "description": "optional; default from service config if \u003c=0",
                    "type": "integer"
                },
                "memory_mb": {
                    "description": "optional; default from service config if \u003c=0",
                    "type": "integer"
                },
                "source_vm_name": {
                    "description": "required; name of existing VM in libvirt to clone from",
                    "type": "string"
                },
                "vm_name": {
                    "description": "optional; generated if empty",
                    "type": "string"
                }
            }
        },
        "virsh-sandbox_internal_rest.createSandboxResponse": {
            "type": "object",
            "properties": {
                "sandbox": {
                    "$ref": "#/definitions/virsh-sandbox_internal_store.Sandbox"
                }
            }
        },
...
    },
...
}